LayoutTests/fast/canvas/canvas-drawImage-detached-leak-expected.txt

Test drawImage() to a detached canvas won't leak the drawn images.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.remoteImagesCountForTesting() is 0
PASS successfullyParsed is true

TEST COMPLETE
